refactor(storage): split SeaORM team repository for GH727
